pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "video_rotation_360"
version = "0.1.0"
description = "A python package for turning rotating footage into 360-degree video."
authors = [
    { name="Jeb Nicholson", email="nicje229@student.otago.ac.nz" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch",
		"torchvision",
		"torchaudio",
		"pillow",
		"numpy",
		"opencv-python", 
		"line-profiler",
		"spatialmedia @ git+https://github.com/google/spatial-media.git",
		"av",
		"matplotlib",
		"scipy",
		"shapely",
		"progressbar2",
		"ngdsac_horizon @ git+https://github.com/jeb5/ngdsac_horizon.git"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 130
indent-width = 2